refactor: MemberResponseDTO 평균별점 형식 수정

+++ back-end/src/main/java/com/studymate/backend/member/dto/MemberResponse.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 
 @AllArgsConstructor
 @Builder
@@ -23,7 +25,7 @@ public class MemberResponse {
     private String publicRelations;
     private String job;
     private int heart;
-    private double starAverage;
+    private BigDecimal starAverage;
     private int solved;
     private int matchingCount;
 }
